Seventh task in second part done and update sixth task.
